model: Make add word changes persistent in letter.properties file
Modify letter.properties file also if a word is added so that next
launch of application can reload it.

Remember that old letter.properties is in bin/ and that needs to
be replaced. Or project needs to be rebuilt for new file to take
affect.

Loading words from database is scheduled for next iteration now.